refactor(tests): share makeSteamIds via Pest.php
