+++ tests/Feature/Http/Requests/ISteamUser/GetPlayerSummariesRequestTest.php
@@ -21,19 +21,6 @@ function summariesConnector(): SteamConnector
     return new SteamConnector(new SteamConfig('test-key'));
 }
 
-/**
- * @return list<SteamId>
- */
-function makeSteamIds(int $count): array
-{
-    $base = 76561198000000000;
-
-    return array_map(
-        static fn (int $offset): SteamId => SteamId::fromSteamId64((string) ($base + $offset)),
-        range(0, $count - 1),
-    );
-}
-
 test('PlayerSummary defaults profilestate and personastate when keys absent', function (): void {
     $summary = PlayerSummary::fromArray([
         'steamid' => '76561198000000000',
